requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.2.2
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
Pillow>=10.0.0
drf-nested-routers>=0.93.4
django-cors-headers>=4.3.0
